Move Completions structure definition into completions module
